pyproject.toml
[project]
name = "web-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "lxml>=6.0.2",
    "pyyaml>=6.0.3",
    "semantic-scholar-api>=0.2.0",
    "tenacity>=9.1.2",
]
